Adding id's to questions and question options
These are required to be able to map student responses over time as the curriculum changes. If the exact wording of an option changes, we can generally compare students between versions of curriculum and the IDs shouldn't change. If the intent of the question or option changes, the id should be changed to reflect that