Adds specs for ci to run lint and validation
We want a simple way for the team to have some validation of
content before importing. It seems important as we reduce the
overhead to importing (both single course and from a branch). I
hope that along with CI running the lint and validation, the team
starts to run it locally as well.